% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHTMLdir.R, R/checkPDFdir.R, R/checkdir.R,
%   R/doc-checkdirs.R
\name{checkHTMLdir}
\alias{checkHTMLdir}
\alias{checkPDFdir}
\alias{checkdir}
\alias{checkdirs}
\title{Extract statistics from folders with PDF/HTML articles and recalculate p-values}
\usage{
checkHTMLdir(dir, subdir = TRUE, extension = TRUE, ...)

checkPDFdir(dir, subdir = TRUE, ...)

checkdir(dir, subdir = TRUE, ...)
}
\arguments{
\item{dir}{String indicating the directory to be used. If left empty, a window 
will pop up from which you can choose a directory.}

\item{subdir}{Logical. Indicates whether you also want to check subfolders. 
Defaults to TRUE.}

\item{extension}{(Only relevant for \code{checkHTMLdir()}) If TRUE (default), only files with a `.html` or `.htm`
extension are processed. If FALSE, all files in the specified directory
(and subdirectories if `subdir = TRUE`) are included, regardless of extension.
Setting this to FALSE may cause errors if non-HTML files are present.}

\item{...}{Arguments sent to \code{statcheck}.}
}
\value{
A \code{statcheck} data frame with the extracted statistics. See 
  \code{\link{statcheck}} for details.
}
\description{
These functions search for NHST results in all PDF and/or HTML articles in a 
certain folder and send the extracted statistics to \code{statcheck}.
}
\details{
On macOS, the file selection dialog can sometimes be unreliable or fail to open
when using `tk_choose.files()`. If you encounter this issue, it is recommended
to manually type the filename or path instead of using the dialog.
}
\note{
This package was primarily tested on Windows. Mac users may experience issues
with multi-file selection dialogs.
}
\examples{
\dontrun{
# Scan a folder of PDFs
checkPDFdir("path/to/pdfs")

# Scan a folder of HTML articles
checkHTMLdir("path/to/html")

# Scan a folder of mixed PDFs/HTML
checkdir("path/to/articles")
}
}
