\name{identify.statcheck}
\alias{identify.statcheck}
\title{identify statcheck}

\usage{\method{identify}{statcheck}(x, ...)}
\arguments{
  \item{x}{
}
  \item{\dots}{
}
}



\author{Sacha Epskamp <mail@sachaepskamp.com> & Michele B. Nuijten <m.b.nuijten@uvt.nl>}





