\name{r2t}
\alias{r2t}
\title{Transform r values into t values}
\description{Function to transform r values into t values by use of raw r and degrees of freedom.}
\usage{r2t(r, df)}
\arguments{
  \item{r}{Raw correlation value}
  \item{df}{Degrees of freedom (N-1)}
}



\author{Sacha Epskamp <mail@sachaepskamp.com> & Michele B. Nuijten <m.b.nuijten@uvt.nl>}





