% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFimport.R, R/checkdir.R, R/doc-checkdirs.R,
%   R/htmlImport.R
\name{checkPDFdir}
\alias{checkPDFdir}
\alias{checkdir}
\alias{checkdirs}
\alias{checkHTMLdir}
\title{Extract statistics from folders with PDF/HTML articles and recalculate 
p-values}
\usage{
checkPDFdir(dir, subdir = TRUE, ...)

checkdir(dir, subdir = TRUE, ...)

checkHTMLdir(dir, subdir = TRUE, extension = TRUE, ...)
}
\arguments{
\item{dir}{String indicating the directory to be used. If this is left empty, 
a window will pop up from which you can choose a directory.}

\item{subdir}{Logical. Indicates whether you also want to check subfolders. 
Defaults to TRUE}

\item{...}{Arguments sent to \code{statcheck}.}

\item{extension}{Logical. Indicates whether the HTML extension should be 
checked. Defaults to TRUE}
}
\value{
A statcheck data frame with the extracted statistics. See 
\code{\link{statcheck}} for details.
}
\description{
These functions search for NHST results in all PDF and/or HTML articles in a 
certain folder and send the extracted statistics to \code{statcheck}.
}
