% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFimport.R, R/doc-checkfiles.R, R/htmlImport.R
\name{checkPDF}
\alias{checkPDF}
\alias{checkfiles}
\alias{checkHTML}
\title{Extract statistics from PDF/HTML articles and recalculate p-values}
\usage{
checkPDF(files, ...)

checkHTML(files, ...)
}
\arguments{
\item{files}{Vector of strings containing file paths to HTML files to check.}

\item{...}{Arguments sent to \code{statcheck}.}
}
\value{
A statcheck data frame with the extracted statistics. See 
\code{\link{statcheck}} for details.
}
\description{
These functions search for NHST results in PDF and/or HTML articles and send 
the extracted statistics to \code{statcheck}.
}
