\name{statcheck-package}
\alias{statcheck-package}
\alias{statcheck}
\docType{package}
\title{Extract statistics from articles and recompute p values}
\description{Extract statistics from articles and recompute p values.}
\details{
\tabular{ll}{Package: \tab statcheck\cr
Type: \tab Package\cr
Title: \tab Extract statistics from articles and recompute p values\cr
Version: \tab 0.1.0\cr
Date: \tab 2013-01-04\cr
Author: \tab Sacha Epskamp <mail@sachaepskamp.com> & Michele B. Nuijten <m.b.nuijten@uvt.nl>\cr
Maintainer: \tab Michele B. Nuijten <m.b.nuijten@uvt.nl>\cr
Depends: \tab 
R (>= 2.14.2)\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
ByteCompile: \tab yes\cr
Imports: \tab plyr\cr}
}
\author{Sacha Epskamp <mail@sachaepskamp.com> & Michele B. Nuijten <m.b.nuijten@uvt.nl>}

\keyword{ package }


