\name{statcheck-package}

\alias{statcheck-package}

\docType{package}

\title{Extract statistics from articles and recompute p values}

\description{Extract statistics from articles and recompute p values.}

\details{

\tabular{ll}{Package: \tab statcheck\cr

Type: \tab Package\cr

Title: \tab Extract statistics from articles and recompute p values\cr

Version: \tab 1.0.0\cr

Date: \tab 2014-11-15\cr

Author: \tab Sacha Epskamp <mail@sachaepskamp.com> & Michele B. Nuijten

<m.b.nuijten@uvt.nl>\cr

Maintainer: \tab Michele B. Nuijten <m.b.nuijten@uvt.nl>\cr

Depends: \tab 

R (>= 2.14.2),

plyr\cr

License: \tab GPL-2\cr

LazyLoad: \tab yes\cr

ByteCompile: \tab yes\cr}

}

\author{Sacha Epskamp <mail@sachaepskamp.com> & Michele B. Nuijten

<m.b.nuijten@uvt.nl>}



\keyword{ package }





