\name{statcheck}
\alias{statcheck}
\title{statcheck}

\usage{statcheck(x, stat = c("t", "F", "cor", "chisq", "Z", "Wald"), 
    OneTailedTests = FALSE, alpha = 0.05, OneTailedTxt = FALSE)}
\arguments{
  \item{x}{
}
  \item{stat}{
}
  \item{OneTailedTests}{
}
  \item{alpha}{
}
  \item{OneTailedTxt}{
}
}



\author{Sacha Epskamp <mail@sachaepskamp.com> & Michele B. Nuijten <m.b.nuijten@uvt.nl>}





