\name{summary.statcheck}
\alias{summary.statcheck}
\title{summary statcheck}

\usage{\method{summary}{statcheck}(object, ...)}
\arguments{
  \item{object}{
}
  \item{\dots}{
}
}



\author{Sacha Epskamp <mail@sachaepskamp.com> & Michele B. Nuijten <m.b.nuijten@uvt.nl>}





